Refactor task queues
