+++ src/game/entities/objects/bullets/bullet.ts
@@ -22,7 +22,6 @@ class Bullet extends Entity {
     if(this.groupId !== other.groupId){
       this.attack(other);
       this.remove();
-      console.log('delete ', this.id);
     }
   }
